Get rid of duplicated forms
