+++ src/components/ConanModal/ConanModalForm.tsx
@@ -3,12 +3,16 @@ import React, { useState } from "react";
 import { connect } from "react-redux";
 import Form from "react-bootstrap/Form";
 import { Field, reduxForm, formValueSelector } from "redux-form";
-import DiffLadder from "../DiffLadder/DiffLadder";
-import "../2d20/form/form.css";
+import DiffLadder from "../2d20/DiffLadder/DiffLadder";
+import "../2d20/form-2d20/form-2d20.css";
 import DiceRow from "../2d20/dice-row/dice-row";
-import { focusLabel, tnLabel, untrainedTestLabel } from "../2d20/labels";
+import {
+  focusLabel,
+  tnLabel,
+  untrainedTestLabel,
+} from "../2d20/form-2d20/labels";
 import Fortune from "../2d20/fortune/fortune";
-import { RenderCheckbox, renderInput } from "../2d20/form/form";
+import { RenderCheckbox, renderInput } from "../2d20/form-2d20/form-utils";
 import Assistance from "../2d20/assistance/assistance";
 
 function ConanModalForm({
@@ -44,10 +48,6 @@ function ConanModalForm({
         change("fortune", "0");
       }
     }
-  };
-
-  const handleAssistanceDiceChange = (diceAmount: string) => {
-    change("assistanceDice", diceAmount);
   };
 
   const handleFocusChange = (focusValue: string) => {
@@ -116,55 +116,6 @@ function ConanModalForm({
     </Form>
   );
 }
-
-interface ErrorPropTypes {
-  tn?: string;
-  focus?: string;
-  assistanceFocus?: string;
-  assistanceTn?: string;
-}
-
-const validate = (values: any) => {
-  const errors: ErrorPropTypes = {};
-  const { tn, focus, assistanceTn, assistanceFocus } = values;
-  const tnNumber = parseInt(tn, 10);
-  const focusNumber = parseInt(focus, 10);
-  const assistanceTnNumber = parseInt(assistanceTn, 10);
-  const assistanceFocusNumber = parseInt(assistanceFocus, 10);
-
-  if (!tn || !tn.trim()) {
-    errors.tn = "Target Number cannot be empty";
-  }
-  if (tnNumber < 0) {
-    errors.tn = "Target Number must be equal or greater than 0";
-  }
-  if (tnNumber > 20) {
-    errors.tn = "Target Number must be less than 20";
-  }
-  if (isNaN(tnNumber)) {
-    errors.tn = "Target Number must be a valid number";
-  }
-  if (focus && isNaN(focusNumber)) {
-    errors.focus = "Focus must be a valid number";
-  }
-  if (focus && focusNumber > 5) {
-    errors.focus = "Focus must be less than 5";
-  }
-  // Assistance values
-  if (assistanceFocus && isNaN(assistanceFocusNumber)) {
-    errors.assistanceFocus = "Focus must be a valid number";
-  }
-  if (assistanceFocus && assistanceFocusNumber > 5) {
-    errors.assistanceFocus = "Focus must be less than 5";
-  }
-  if (assistanceTn && isNaN(assistanceTnNumber)) {
-    errors.assistanceTn = "TN must be a valid number";
-  }
-  if (assistanceTnNumber && assistanceTnNumber > 20) {
-    errors.assistanceTn = "Target Number must be less than 20";
-  }
-  return errors;
-};
 
 const form = "ConanModalForm";
